r-gen-app/requirements.txt
fastapi==0.135.3
uvicorn[standard]==0.43.0
httpx==0.28.1
pandas==3.0.2
openpyxl==3.1.5
xlrd==2.0.2
python-dotenv==1.2.2
python-multipart==0.0.23
numpy==2.4.4
